+++ Cartfile.resolved
@@ -15,7 +15,7 @@ github "wireapp/protobuf-objc" "1.9.14"
 github "wireapp/swift-protobuf" "1.5.0"
 github "wireapp/wire-ios-canvas" "10.0.0"
 github "wireapp/wire-ios-cryptobox" "18.0.0"
-github "wireapp/wire-ios-data-model" "feature/legalhold-expose-clients"
+github "wireapp/wire-ios-data-model" "feature/legalhold"
 github "wireapp/wire-ios-images" "26.0.0"
 github "wireapp/wire-ios-link-preview" "22.0.0"
 github "wireapp/wire-ios-protos" "20.0.0"
